feat: WearableRepresentation use ajv validations + move tests to /item
