+++ test/platform/item/body-shape.spec.ts
