+++ test/platform/item/locale.spec.ts
