+++ test/platform/item/metrics.spec.ts
